+++ gframe/image_manager.h
@@ -3,6 +3,7 @@
 
 #include "config.h"
 #include <path.h>
+#include <rect.h>
 #include <unordered_map>
 #include <atomic>
 #include <future>
@@ -34,7 +35,7 @@ public:
 private:
 	using image_path = std::pair<irr::video::IImage*, path_string>;
 	using loading_map = std::map<int, std::future<image_path>>;
-	using chrono_time = unsigned long long;
+	using chrono_time = uint64_t;
 	using texture_map = std::unordered_map<uint32_t, irr::video::ITexture*>;
 public:
 	ImageManager() {
